refactor(dashboard): move api utils out of `components`
